refactor: application DI registration changes
