+++ Source/Endpoints/Iwentys.Endpoints.Api/Areas/Identity/IdentityHostingStartup.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Identity.UI;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.AspNetCore.Hosting;
 
 [assembly: HostingStartup(typeof(Iwentys.Endpoints.Api.Areas.Identity.IdentityHostingStartup))]
 namespace Iwentys.Endpoints.Api.Areas.Identity
