+++ Source/Endpoints/Iwentys.Endpoints.Api/Areas/Identity/Pages/Account/Login.cshtml.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
-using Iwentys.Infrastructure.Application.Authorization;
+using Iwentys.Endpoints.Api.Authorization;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Identity;
