+++ Source/Endpoints/Iwentys.Endpoints.Api/Authorization/ApplicationUser.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace Iwentys.Endpoints.Api.Authorization
+{
+    public class ApplicationUser : IdentityUser
+    {
+    }
+}
