+++ Source/Endpoints/Iwentys.Endpoints.Api/Pages/Shared/_LoginPartial.cshtml
@@ -1,5 +1,4 @@
-﻿@using Microsoft.AspNetCore.Identity
-@using Iwentys.Infrastructure.Application.Authorization
+﻿@using Iwentys.Endpoints.Api.Authorization
 
 @inject SignInManager<ApplicationUser> SignInManager
 @inject UserManager<ApplicationUser> UserManager
