+++ Source/Infrastructure/Iwentys.Infrastructure.Application/AuthorizedUser.cs
@@ -1,0 +1,18 @@
+﻿namespace Iwentys.Infrastructure.Application.Authorization
+{
+    public sealed class AuthorizedUser
+    {
+        private AuthorizedUser(int id)
+        {
+            Id = id;
+        }
+
+        public int Id { get; set; }
+
+        //TODO: think about. Only for debug propose
+        public static AuthorizedUser DebugAuth(int id)
+        {
+            return new AuthorizedUser(id);
+        }
+    }
+}
